update makefile to make it work on kiwi
