--- coreir_lib.h
@@ -1,10 +1,0 @@
-#pragma once
-
-#ifdef COREIR
-
-#include "coreir/common-macros.h"
-#include "coreir-c/ctypes.h"
-#include "coreir/libs/commonlib.h"
-
-COREIR_GEN_CPP_API_DECLARATION_FOR_LIBRARY(cwlib);
-#endif
